no message
git-svn-id: https://svn.code.sf.net/p/khc/svn/trunk/core@481 ab1598f5-0ef3-491f-8894-8c1f2390833e

+++ kh_lib/gui_window/word_df_freq.pm
@@ -1,4 +1,4 @@
-package gui_window::word_freq;
+package gui_window::word_df_freq;
 use base qw(gui_window);
 
 use strict;
@@ -11,7 +11,7 @@ sub _new{
 	my $wmw= $self->{win_obj};
 	#$self->{win_obj} = $wmw;
 	#$wmw->focus;
-	$wmw->title($self->gui_jchar('出現回数：分布'));
+	$wmw->title($self->gui_jchar('文書数：分布'));
 	
 	$wmw->Label(
 		-text => $self->gui_jchar('■記述統計'),
@@ -81,12 +81,21 @@ sub _new{
 		)->pack(-side => 'left');
 	}
 
-	$wmw->Button(
-		-text => $self->gui_jchar('再計算'),
-		-font => "TKFN",
-		-borderwidth => '1',
-		-command => sub{ $mw->after(10,sub {$self->count;});} 
-	)->pack(-side => 'left',-padx => 5);
+	$wmw->Label(
+		-text => $self->gui_jchar('  集計単位：'),
+		-font => "TKFN"
+	)->pack(-side => 'left');
+
+	my %pack = (
+			#-anchor => 'e',
+			#-pady   => 1,
+			-side   => 'left'
+	);
+	$self->{tani_obj} = gui_widget::tani->open(
+		parent  => $wmw,
+		pack    => \%pack,
+		command => sub {$self->count;},
+	);
 
 	$wmw->Button(
 		-text => $self->gui_jchar('閉じる'),
@@ -95,8 +104,6 @@ sub _new{
 		-command => sub{ $mw->after(10,sub {$self->close;});} 
 	)->pack(-side => 'right');
 
-
-
 	$self->{list1} = $lis1;
 	$self->{list2} = $lis2;
 	return $self;
@@ -104,7 +111,8 @@ sub _new{
 
 sub count{
 	my $self = shift;
-	my ($r1, $r2) = mysql_words->freq_of_f;
+	return 0 unless $self->{tani_obj};
+	my ($r1, $r2) = mysql_words->freq_of_df($self->{tani_obj}->tani);
 	
 	# 記述統計
 	$self->list1->delete('all');
@@ -158,7 +166,7 @@ sub count{
 	$rcmd .= "), nrow=$row, ncol=3, byrow=TRUE)";
 	$self->{rcmd} = $rcmd;
 	
-	if ($::main_gui->if_opened('w_word_freq_plot')){
+	if ($::main_gui->if_opened('w_word_df_freq_plot')){
 		$self->plot;
 		$self->{win_obj}->focus;
 	}
@@ -171,9 +179,9 @@ sub plot{
 	my $self = shift;
 	return 0 unless $::config_obj->R;
 	
-	my $path1 = $::project_obj->dir_CoderData.'words_TF_freq1.bmp';
-	my $path2 = $::project_obj->dir_CoderData.'words_TF_freq2.bmp';
-	my $path3 = $::project_obj->dir_CoderData.'words_TF_freq3.bmp';
+	my $path1 = $::project_obj->dir_CoderData.'words_DF_freq1.bmp';
+	my $path2 = $::project_obj->dir_CoderData.'words_DF_freq2.bmp';
+	my $path3 = $::project_obj->dir_CoderData.'words_DF_freq3.bmp';
 	$path1 =~ tr/\\/\//;
 	$path2 =~ tr/\\/\//;
 	$path3 =~ tr/\\/\//;
@@ -183,23 +191,23 @@ sub plot{
 	$::config_obj->R->send($self->{rcmd});
 	# 通常
 	$::config_obj->R->send("bmp(\"$path1\")");
-	$::config_obj->R->send('matplot(hage[,1],hage[,3],type="b",lty=1,pch=1,ylab="Freqency", xlab="TF")');
+	$::config_obj->R->send('matplot(hage[,1],hage[,3],type="b",lty=1,pch=1,ylab="Freqency", xlab="DF")');
 	$::config_obj->R->send('dev.off()');
 	# x軸を対数に
 	$::config_obj->R->send("bmp(\"$path2\")");
-	$::config_obj->R->send('matplot(hage[,1],hage[,3],type="b",lty=1,pch=1,log="x",ylab="Freqency", xlab="TF")');
+	$::config_obj->R->send('matplot(hage[,1],hage[,3],type="b",lty=1,pch=1,log="x",ylab="Freqency", xlab="DF")');
 	$::config_obj->R->send('dev.off()');
 	# xy軸を対数に
 	$::config_obj->R->send("bmp(\"$path3\")");
-	$::config_obj->R->send('plot(hage[,1],hage[,3],log="xy",ylab="Freqency", xlab="TF")');
+	$::config_obj->R->send('plot(hage[,1],hage[,3],log="xy",ylab="Freqency", xlab="DF")');
 	$::config_obj->R->send('dev.off()');
 	$::config_obj->R->unlock;
 	$::config_obj->R->output_chk(1);
 	
-	if ($::main_gui->if_opened('w_word_freq_plot')){
-		$::main_gui->get('w_word_freq_plot')->renew;
+	if ($::main_gui->if_opened('w_word_df_freq_plot')){
+		$::main_gui->get('w_word_df_freq_plot')->renew;
 	} else {
-		gui_window::word_freq_plot->open(
+		gui_window::word_df_freq_plot->open(
 			images => [$path1,$path2,$path3]
 		);
 	}
@@ -217,7 +225,7 @@ sub list1{
 	return $self->{list1};
 }
 sub win_name{
-	return 'w_word_freq';
+	return 'w_word_df_freq';
 }
 
 1;
